[mod/#75]AddSportsNavigation 패키지 이동
