+++ app/src/main/java/com/smashing/app/presentation/addsports/navigation/AddSportsNavigation.kt
@@ -1,0 +1,30 @@
+package com.smashing.app.presentation.addsports.navigation
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavOptions
+import androidx.navigation.compose.composable
+import com.smashing.app.core.common.navigation.Route
+import com.smashing.app.presentation.addsports.AddSportsRoute
+import kotlinx.serialization.Serializable
+
+
+fun NavController.navigateToAddSports(
+    navOptions: NavOptions? = null,
+) = navigate(AddSports, navOptions)
+
+fun NavGraphBuilder.addSportsGraph(
+    navigateUp: () -> Unit,
+) {
+    composable<AddSports> {
+        AddSportsRoute(
+            modifier = Modifier.fillMaxSize(),
+            navigateToUser = navigateUp,
+        )
+    }
+}
+
+@Serializable
+data object AddSports : Route
