+++ app/src/main/java/com/smashing/app/presentation/main/MainScreen.kt
@@ -12,6 +12,8 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.compose.NavHost
 import androidx.navigation.navOptions
 import com.smashing.app.core.designsystem.theme.SmashingTheme
+import com.smashing.app.presentation.addsports.navigation.addSportsGraph
+import com.smashing.app.presentation.addsports.navigation.navigateToAddSports
 import com.smashing.app.presentation.home.navigation.homeGraph
 import com.smashing.app.presentation.home.navigation.navigateToHome
 import com.smashing.app.presentation.login.navigation.Login
@@ -21,7 +23,6 @@ import com.smashing.app.presentation.matching.navigation.Matching
 import com.smashing.app.presentation.matching.navigation.matchingGraph
 import com.smashing.app.presentation.matching.navigation.navigateToMatching
 import com.smashing.app.presentation.notice.navigation.noticeGraph
-import com.smashing.app.presentation.profile.navigation.navigateToAddSports
 import com.smashing.app.presentation.profile.navigation.navigateToReview
 import com.smashing.app.presentation.profile.navigation.profileGraph
 import com.smashing.app.presentation.ranking.navigation.rankingGraph
@@ -180,6 +181,9 @@ private fun MainNavHost(
         tierInfoGraph(
             innerPadding = innerPadding,
             navController = appState.navController,
+        )
+        addSportsGraph(
+            navigateUp = appState.navController::navigateUp,
         )
     }
 }
